+++ app/src/main/java/reservant_mobile/ui/components/MenuComponents.kt
@@ -9,10 +9,13 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.DeleteForever
 import androidx.compose.material.icons.filled.Edit
 import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.MaterialTheme
@@ -398,4 +401,23 @@ fun AddMenuButton(
     MyFloatingActionButton(
         onClick = { showAddDialog.value = true }
     )
+}
+
+@Composable
+fun MenuTypeButton(
+    modifier: Modifier = Modifier,
+    menuType: String,
+    onMenuClick: () -> Unit
+) {
+    Button(
+        onClick = { onMenuClick() },
+        shape = RoundedCornerShape(50),
+        colors = ButtonDefaults.buttonColors(
+            containerColor = MaterialTheme.colorScheme.secondaryContainer,
+            contentColor = MaterialTheme.colorScheme.onSecondaryContainer
+        ),
+        modifier = modifier.padding(4.dp)
+    ) {
+        Text(menuType)
+    }
 }
